+++ src/utils.ts
@@ -10,7 +10,6 @@ function getDaysUntil(dt1: number, dt2: number): number {
   // 締め切りまでの日数を計算します
   let diff = (dt2 - dt1) / 1000;
   diff /= 3600 * 24;
-  // if (diff < 0) diff = 9999;
   return diff;
 }
 
@@ -19,23 +18,6 @@ function getTimeRemain(_remainTime: number): [number, number, number] {
   const hours = Math.floor((_remainTime - day * 3600 * 24) / 3600);
   const minutes = Math.floor((_remainTime - (day * 3600 * 24 + hours * 3600)) / 60);
   return [day, hours, minutes];
-}
-
-// TODO: dom.tsに移動したい
-function createElem(tag: any, dict?: { [key: string]: any }): any {
-  const elem = document.createElement(tag);
-  for (const key in dict) {
-    // @ts-ignore
-    elem[key] = dict[key];
-  }
-  return elem;
-}
-
-function appendChildAll(to: HTMLElement, arr: Array<any>): HTMLElement {
-  for (const obj in arr) {
-    to.appendChild(arr[obj]);
-  }
-  return to;
 }
 
 function createLectureIDMap(lectureIDList: Array<LectureInfo>): Map<string, string> {
@@ -81,7 +63,6 @@ function updateIsReadFlag(kadaiList: Array<Kadai>): void {
         updatedKadaiList.push(kadai);
       }
     }
-    console.log("フラグをへし折りました！")
     saveToStorage("TSkadaiList", updatedKadaiList);
   }
 }
@@ -137,7 +118,6 @@ function compareAndMergeKadaiList(oldKadaiList: Array<Kadai>, newKadaiList: Arra
         // もしなければ新規課題なので未読フラグを立てる
         if (q === -1) {
           isRead = false;
-          console.log("cannnot find");
           mergedKadaiEntries.push(newKadaiEntry);
         } else {
           const entry = new KadaiEntry(
@@ -149,8 +129,8 @@ function compareAndMergeKadaiList(oldKadaiList: Array<Kadai>, newKadaiList: Arra
             newKadaiEntry.assignmentDetail
           );
           entry.kadaiPage = newKadaiEntry.kadaiPage;
-          mergedKadaiEntries.push(entry)
-        };
+          mergedKadaiEntries.push(entry);
+        }
       }
       // 未読フラグ部分を変更してマージ
       mergedKadaiEntries.sort((a, b)=>{return a.dueDateTimestamp - b.dueDateTimestamp});
@@ -182,7 +162,7 @@ function useCache(fetchedTime: number): boolean{
   return (nowTime - fetchedTime) / 1000 > cacheInterval;
 }
 
-function genUniqueStr() {
+function genUniqueStr(): string {
   return "m" + new Date().getTime().toString(16) + Math.floor(123456 * Math.random()).toString(16);
 }
 
@@ -190,8 +170,6 @@ function genUniqueStr() {
 export {
   getDaysUntil,
   getTimeRemain,
-  createElem,
-  appendChildAll,
   createLectureIDMap,
   isLoggedIn,
   miniPandAReady,
